moved ProductRepository to src/Repository
